feat: Add configuration to allow toggling the plugin on and off
Previously, the plugin will just load itself up and run as soon as it is dropped into the plugins
folder. This change allows the end user to turn off the plugin without uninstalling it or even
disable certain metrics they do not want. But the default is still to keep all metrics enabled.
This also makes it possible to do the same for when the plugin can be installed via PyPI.