Reduce log level of blacklisted domains from WARNING to INFO
Fixes #49, in which I accidentally reduced the log level of the wrong message.

This commit reduces the log level for the message "Found blacklisted domain" in `patch_requests.py` from `WARNING` to `INFO` because each warning log sends a message to Sentry. For my team, this becomes over 17,000 warnings every day, which is excessive for this warning that is largely unactionable.